Change the order of delete cascade for `spiral'
A crash of `folsom_metrics_histogram_ets` leaves spirals broken.
The spiral ets table goes away, so it can not be updated, nor
can it be deleted and re-created due to the delete order of
metric table, spiral table, folsom table. Have changed the order
to be more like that of `histogram` so deletes will succeed in
the event of a missing `spiral` ets table.